+++ tests/views/test_projects_view.py
@@ -28,3 +28,17 @@ class TestProjectView(BaseTestView):
         'customer_id': 'c2034c1b-0a40-4b84-9ace-54b958f64ed4',
         'title': 'Other Name'
     }
+
+    def test_put_invalid_asset_tyoe(self, app, obj_payload):
+        """Asset type should match one of the possible values."""
+        payload = obj_payload
+        obj_id = payload['id']
+        payload['asset_types'] = ['Foobar']
+        request = app.put_json('{base}/{id}'.format(base=self.base_path, id=obj_id),
+                               payload, headers=self.headers, status=400)
+        result = request.json
+        error = result['errors'][0]
+        assert result['status'] == 'error'
+        assert error['name'] == 'asset_types'
+        assert error['location'] == 'body'
+        assert 'Invalid type of asset' in error['description']